refactor: enhance build process and update recipient parsing
- Updated the build script in package.json to clean the lib directory before building.
- Refactored recipient parsing function name from `parseRecipent` to `parseRecipient` for consistency.
- Added unit tests for recipient parsing to ensure correct functionality.
- Removed outdated integration tests to streamline the test suite.
